+++ Data/Entities/User/UserMood.cs
@@ -37,5 +37,5 @@ public class UserMood : IScore
 
     public int? ProratedScore => Items.Sum();
 
-    public int? AverageScore => Items.Sum() / Items.Count;
+    public double? AverageScore => Items.Sum() / (double)Items.Count;
 }
